+++ CHANGELOG.md
@@ -79,3 +79,8 @@ All notable changes to this service will be documented in this file.
 ### Notes
 - Added new Role assignment endpoint
 - New endpoint for retrieving tenant-wide applications
+- Added RG Visits Test env appsettings
+
+## [1.3.1] - Public Beta
+### Notes
+- Added the ability to search for applications by reference on the dashboard